Update tests for preserved aspect ratios
Now that we default to preserving image aspect ratios, these tests
actually end up showing very thin borders of background colour. This
is because, being embedded in a larger app, they aren't quite the size
we think. So preserving the correct aspect ratio leaves miniscule
letter/pillar-box boarders.

Probably the best behaviour is simply to fill the window and allow the
aspect ratio to be changed very marginally. It's not remotely visible,
and seems more desirable in these cases.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>