Set framerate ranges for preview and still configurations
We set preview to run by default from 12fps to the fastest possible,
and still capture to anything the sensor will allow. This is
reasonable in itself, but also means you can switch away to a video
mode and then back without typically getting stuck at the fixed video
framerate of 30fps.

This was all customisable previously (and still is), but the default
behaviours if applications don't pay attention are probably better
like this.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>