Add YUV420 support to the QPicamera2 widget
This is the non-OpenGL accelerated version which previously could not
display YUV420 format images. It can do so now through the use of
OpenCV which provides a much faster YUV to RGB conversion function
than we can write easily using numpy. (But be under no illusions that
it still costs quite a bit of CPU!)

If OpenCV is not available, an error is only thrown if you try to
display a YUV420 image, everything else will work as before.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>