Linting using flake8, include extra docstrings
This adds a GitHub action for flake8 linting along with an initial
pass over the code base, adding docstrings and running isort on the
Python files. Compilation of libcamera is also performed to allow
flake8 to import it.

This is a first commit in a series. There will be at least one more to
add further missing docstrings and actually to flag flake8 complaints
as errors.

Signed-off-by: Chris Richardson <christopher.richardson@raspberrypi.org>