Add extra default parameters to the start function
You can now specify a configuration if you want to and you can also
ask it start the preview (avoiding the need to call start_preview
before). The event_loop parameter is gone, replaced by show_preview
which would defaults to False giving the previous behaviour, but None
can be passed to avoid starting an event_loop altogether.

start_preview now supports "True" as a parameter which will perform a
(simple-minded) detection of which preview to start. In most simple
circumstances it should to the right thing, but is not guaranteed.

One test that was using the old event_loop parameter has been
updated.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>