Some internal refactoring using futures
We now use a future object rather than events to signal when an
operation is complete. This makes handling the result value and errors
tidier.

The behaviour is changed so that you have to call wait() to get the
result of the asynchronous operation, obviously you can pass a
signal_function to notify you when it's safe to call wait without it
blocking. All the old async_result, async_error etc. fields have been
removed.

The default behaviour is now for funtions to block if you don't pass a
signal_function, but to return immediately if you do (though this can
be overridden).

The apps/app_capture.py example has been removed because it's much
tidier (even more so now) to pass in the slot signal function and do
things properly (as apps/app_capture2.py does).

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>