Set logging appropriately for all the test scripts
The revised logging scheme has the behaviour that all warning and
error messages get printed by default, but without the "WARN" or
"ERROR" tag at the start, meaning that the automated testing might
miss them.

So for all our formal "test scripts" in the tests directory, let's set
the logging up so that those tags appear. The default call to
set_logging will do this.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>