Change the signature of the wait() function
You now have to say what you're waiting for. This is a breaking API
change (although only in a relatively minor way) affecting only people
who make asynchronous calls. The change actually does nothing
immediately, but will be useful in subsequent commits.

Mostly the ramifications affect the Qt examples. Because you're not
always sure whether signals cause slot functions to run
synchronously or not, we have to make sure that we pass the "job" all
the way through Qt and on to the application's callback.

So although the changes are small, quite a few files are touched.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>