Cosmetic changes in handling previews
The have_event_loop field is renamed to _event_loop_running. The
"asynchronous" property is removed because it only features in one
place, and we can replace it and add some improved comments.

With these changes we note the the API for starting/stopping preview
windows is as follows:

* The Picamera2 objects calls start_preview to start the preview and
  stop_preview to stop it.

* The preview window calls attach_preview when it starts (in response
  to start_preview), and detach_preview when it stops (in response to
  stop_preview, or spontaneously).

The Picamera2 object's knowledge of the preview window is held only in
the _preview and _event_loop_running fields, and the Picamera2 object
accesses no other information from it.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>