Add switch_mode_and_capture_request
"Old" requests (where the camera configuration has changed since the
request was created) are now handled, so this function is now
potentially useful (the returned request is forcibly "old" in this way
when it is returned).

It does increase the general risk of CMA fragmentation, so should be
used with care.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>