Move the call to render the camera image into process_requests
Previously handle_requests received the request to display as the
return value of process_requests, but now process_requests makes a
call to the new render_request method.

This method is simply the old display code, moved out of the preview's
handle_request method. Calling the render function from within
process_requests allows a bit of tidying, and we can also move the
signalling to the user code after the rendering has been done - which
is how it should be.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>