Add autofocus support
Mostly autofocus can work through the libcamera controls API, but we
add a convenient Picamera2.autofocus_cycle function to start an
autofocus cyle and wait for it to finish.

We also patch up a couple of places where we have to cope with
controls that contain a list of libcamera Rectangles (the 'AfWindow'
control).

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>